Cpp: Removed some gcc compile warnings

+++ Cpp/source/NetworkSupport.cpp
@@ -37,7 +37,7 @@ namespace ops
 
 	uint32_t IPString2Addr(Address_T addr) 
 	{
-		return boost::asio::ip::address_v4::from_string(addr.c_str()).to_ulong();
+		return (uint32_t)boost::asio::ip::address_v4::from_string(addr.c_str()).to_ulong();
 	}
 
 	Address_T IPAddr2String(uint32_t addr)
